Move ChannelInterface and VoidReturningChannelInterface to infrastructure:events and remove its dependency on util.
