+++ infrastructure/events/src/main/java/tech/pegasys/teku/infrastructure/events/EventChannels.java
@@ -19,8 +19,6 @@ import java.util.concurrent.ConcurrentMap;
 import java.util.function.Function;
 import org.hyperledger.besu.plugin.services.MetricsSystem;
 import tech.pegasys.teku.infrastructure.async.AsyncRunner;
-import tech.pegasys.teku.util.channels.ChannelInterface;
-import tech.pegasys.teku.util.channels.VoidReturningChannelInterface;
 
 public class EventChannels {
