+++ infrastructure/events/src/main/java/tech/pegasys/teku/infrastructure/events/VoidReturningChannelInterface.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.util.channels;
+package tech.pegasys.teku.infrastructure.events;
 
 /**
  * Marker interface event channel interfaces which only have {@code void} return types.
