+++ pow/src/main/java/tech/pegasys/teku/pow/api/Eth1EventsChannel.java
@@ -14,10 +14,10 @@
 package tech.pegasys.teku.pow.api;
 
 import org.apache.tuweni.bytes.Bytes32;
+import tech.pegasys.teku.infrastructure.events.VoidReturningChannelInterface;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.pow.event.DepositsFromBlockEvent;
 import tech.pegasys.teku.pow.event.MinGenesisTimeBlockEvent;
-import tech.pegasys.teku.util.channels.VoidReturningChannelInterface;
 
 public interface Eth1EventsChannel extends VoidReturningChannelInterface {
   void onDepositsFromBlock(DepositsFromBlockEvent event);
