+++ storage/api/src/main/java/tech/pegasys/teku/storage/api/Eth1DepositStorageChannel.java
@@ -14,8 +14,8 @@
 package tech.pegasys.teku.storage.api;
 
 import tech.pegasys.teku.infrastructure.async.SafeFuture;
+import tech.pegasys.teku.infrastructure.events.ChannelInterface;
 import tech.pegasys.teku.storage.api.schema.ReplayDepositsResult;
-import tech.pegasys.teku.util.channels.ChannelInterface;
 
 public interface Eth1DepositStorageChannel extends ChannelInterface {
   SafeFuture<ReplayDepositsResult> replayDepositEvents();
