+++ storage/src/main/java/tech/pegasys/teku/storage/api/StorageQueryChannel.java
@@ -22,9 +22,9 @@ import tech.pegasys.teku.datastructures.blocks.SignedBlockAndState;
 import tech.pegasys.teku.datastructures.blocks.SlotAndBlockRoot;
 import tech.pegasys.teku.datastructures.state.BeaconState;
 import tech.pegasys.teku.infrastructure.async.SafeFuture;
+import tech.pegasys.teku.infrastructure.events.ChannelInterface;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.storage.store.StoreBuilder;
-import tech.pegasys.teku.util.channels.ChannelInterface;
 
 public interface StorageQueryChannel extends ChannelInterface {
