+++ storage/src/main/java/tech/pegasys/teku/storage/api/StorageUpdateChannel.java
@@ -14,9 +14,9 @@
 package tech.pegasys.teku.storage.api;
 
 import tech.pegasys.teku.infrastructure.async.SafeFuture;
+import tech.pegasys.teku.infrastructure.events.ChannelInterface;
 import tech.pegasys.teku.storage.events.AnchorPoint;
 import tech.pegasys.teku.storage.events.StorageUpdate;
-import tech.pegasys.teku.util.channels.ChannelInterface;
 
 public interface StorageUpdateChannel extends ChannelInterface {
