+++ util/src/main/java/tech/pegasys/teku/util/time/channels/SlotEventsChannel.java
@@ -13,8 +13,8 @@
 
 package tech.pegasys.teku.util.time.channels;
 
+import tech.pegasys.teku.infrastructure.events.VoidReturningChannelInterface;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
-import tech.pegasys.teku.util.channels.VoidReturningChannelInterface;
 
 public interface SlotEventsChannel extends VoidReturningChannelInterface {
   void onSlot(UInt64 slot);
